fix auth test issues
