--- __tests__/auth/rate-limiting.test.ts
@@ -1,387 +1,0 @@
-import { POST as POST_login } from "@/app/api/auth/login/route";
-import { POST as POST_verify } from "@/app/api/auth/verify/route";
-import { Keypair } from "stellar-sdk";
-import { createAuthRequest } from "../helpers/auth-helpers";
-
-/**
- * Rate limiting tests for authentication endpoints
- * Security-critical: Prevents brute force attacks
- */
-describe("Rate Limiting Tests", () => {
-  let testKeypair: Keypair;
-
-  beforeEach(() => {
-    testKeypair = Keypair.random();
-    jest.clearAllMocks();
-  });
-
-  describe("Login endpoint rate limiting", () => {
-    it("should track and limit excessive login attempts from same public key", async () => {
-      const publicKey = testKeypair.publicKey();
-      const maxAttempts = 10;
-      const responses: Response[] = [];
-
-      // Simulate rapid login attempts
-      for (let i = 0; i < maxAttempts; i++) {
-        const request = new Request("http://localhost/api/auth/login", {
-          method: "POST",
-          body: JSON.stringify({ publicKey }),
-        });
-
-        responses.push(await POST_login(request));
-      }
-
-      // Check if any requests were rate limited
-      const statusCodes = responses.map((r) => r.status);
-      const hasRateLimiting = statusCodes.includes(429);
-
-      // Log for manual verification (implementation needed)
-      if (!hasRateLimiting) {
-        console.warn(
-          "⚠️  Rate limiting not implemented - consider adding rate limiting middleware",
-        );
-      }
-
-      // This test documents expected behavior
-      // Actual implementation should return 429 after threshold
-      expect(responses.length).toBe(maxAttempts);
-    });
-
-    it("should allow normal request rate", async () => {
-      const publicKey = testKeypair.publicKey();
-      const normalAttempts = 3;
-
-      for (let i = 0; i < normalAttempts; i++) {
-        const request = new Request("http://localhost/api/auth/login", {
-          method: "POST",
-          body: JSON.stringify({ publicKey }),
-        });
-
-        const response = await POST_login(request);
-        const data = await response.json();
-
-        // Normal requests should always succeed
-        expect(response.status).toBe(200);
-        expect(data.success).toBe(true);
-
-        // Small delay between requests
-        await new Promise((resolve) => setTimeout(resolve, 100));
-      }
-    });
-
-    it("should have different rate limits per public key", async () => {
-      const keypair1 = Keypair.random();
-      const keypair2 = Keypair.random();
-
-      // Make requests with first keypair
-      for (let i = 0; i < 5; i++) {
-        const request = new Request("http://localhost/api/auth/login", {
-          method: "POST",
-          body: JSON.stringify({ publicKey: keypair1.publicKey() }),
-        });
-        await POST_login(request);
-      }
-
-      // Requests with second keypair should not be affected
-      const request2 = new Request("http://localhost/api/auth/login", {
-        method: "POST",
-        body: JSON.stringify({ publicKey: keypair2.publicKey() }),
-      });
-
-      const response2 = await POST_login(request2);
-      expect(response2.status).toBe(200);
-    });
-  });
-
-  describe("Verification endpoint rate limiting", () => {
-    it("should track failed verification attempts", async () => {
-      // Get a valid challenge first
-      const loginRequest = new Request("http://localhost/api/auth/login", {
-        method: "POST",
-        body: JSON.stringify({ publicKey: testKeypair.publicKey() }),
-      });
-
-      const loginResponse = await POST_login(loginRequest);
-      const loginData = await loginResponse.json();
-      const { nonce, timestamp } = loginData.data;
-
-      // Attempt multiple verifications with invalid signatures
-      const maxAttempts = 10;
-      const responses: Response[] = [];
-
-      for (let i = 0; i < maxAttempts; i++) {
-        const verifyBody = {
-          publicKey: testKeypair.publicKey(),
-          signature: "invalid-signature-" + i,
-          nonce,
-          timestamp,
-        };
-
-        const verifyRequest = createAuthRequest(
-          "http://localhost/api/auth/verify",
-          verifyBody,
-        );
-
-        responses.push(await POST_verify(verifyRequest));
-      }
-
-      // Check results
-      const unauthorizedCount = responses.filter(
-        (r) => r.status === 401,
-      ).length;
-      const rateLimitedCount = responses.filter((r) => r.status === 429).length;
-
-      // Document expected behavior
-      if (rateLimitedCount === 0) {
-        console.warn(
-          "⚠️  Verification rate limiting not implemented - brute force vulnerability",
-        );
-      }
-
-      expect(unauthorizedCount + rateLimitedCount).toBe(maxAttempts);
-    });
-
-    it("should not rate limit successful verifications", async () => {
-      // Successful auth attempts should not be rate limited
-      for (let i = 0; i < 3; i++) {
-        // Get challenge
-        const loginRequest = new Request("http://localhost/api/auth/login", {
-          method: "POST",
-          body: JSON.stringify({ publicKey: testKeypair.publicKey() }),
-        });
-
-        const loginResponse = await POST_login(loginRequest);
-        const loginData = await loginResponse.json();
-        const { nonce, timestamp, message } = loginData.data;
-
-        // Sign and verify
-        const signature = testKeypair
-          .sign(Buffer.from(message, "utf-8"))
-          .toString("base64");
-
-        const verifyBody = {
-          publicKey: testKeypair.publicKey(),
-          signature,
-          nonce,
-          timestamp,
-        };
-
-        const verifyRequest = createAuthRequest(
-          "http://localhost/api/auth/verify",
-          verifyBody,
-        );
-
-        const response = await POST_verify(verifyRequest);
-
-        // Should always succeed
-        expect(response.status).toBe(200);
-
-        await new Promise((resolve) => setTimeout(resolve, 100));
-      }
-    });
-  });
-
-  describe("Rate limit response format", () => {
-    it("should return proper error structure when rate limited", async () => {
-      // This test documents expected response format
-      const expectedResponse = {
-        success: false,
-        error: {
-          code: "RATE_LIMIT_EXCEEDED",
-          message: "Too many requests. Please try again later.",
-          retryAfter: 60, // seconds
-        },
-      };
-
-      // Verify this structure is returned when implementing rate limiting
-      expect(expectedResponse.success).toBe(false);
-      expect(expectedResponse.error.code).toBe("RATE_LIMIT_EXCEEDED");
-      expect(expectedResponse.error).toHaveProperty("retryAfter");
-    });
-
-    it("should include Retry-After header when rate limited", async () => {
-      // Expected header format
-      const expectedHeaders = {
-        "Retry-After": "60",
-        "X-RateLimit-Limit": "10",
-        "X-RateLimit-Remaining": "0",
-        "X-RateLimit-Reset": expect.any(String),
-      };
-
-      // Document expected behavior
-      expect(expectedHeaders).toHaveProperty("Retry-After");
-      expect(expectedHeaders).toHaveProperty("X-RateLimit-Limit");
-    });
-  });
-
-  describe("Distributed rate limiting", () => {
-    it("should handle rate limits across multiple instances", async () => {
-      // In production, rate limiting should work across load-balanced instances
-      // This requires shared state (Redis, database, etc.)
-
-      const publicKey = testKeypair.publicKey();
-
-      // Simulate requests from different "servers"
-      const instance1Requests = Array(5)
-        .fill(null)
-        .map(
-          () =>
-            new Request("http://localhost/api/auth/login", {
-              method: "POST",
-              body: JSON.stringify({ publicKey }),
-              headers: { "X-Instance-ID": "server-1" },
-            }),
-        );
-
-      const instance2Requests = Array(5)
-        .fill(null)
-        .map(
-          () =>
-            new Request("http://localhost/api/auth/login", {
-              method: "POST",
-              body: JSON.stringify({ publicKey }),
-              headers: { "X-Instance-ID": "server-2" },
-            }),
-        );
-
-      // Execute requests
-      const allRequests = [...instance1Requests, ...instance2Requests];
-      const responses = await Promise.all(
-        allRequests.map((req) => POST_login(req)),
-      );
-
-      // Combined requests should trigger rate limiting
-      const statusCodes = responses.map((r) => r.status);
-
-      // Log implementation status
-      if (!statusCodes.includes(429)) {
-        console.warn(
-          "⚠️  Distributed rate limiting not implemented - use Redis or similar",
-        );
-      }
-
-      expect(responses.length).toBe(10);
-    });
-  });
-
-  describe("Rate limit bypass attempts", () => {
-    it("should not be bypassed by changing user agent", async () => {
-      const publicKey = testKeypair.publicKey();
-      const responses: Response[] = [];
-
-      // Try to bypass with different user agents
-      const userAgents = [
-        "Mozilla/5.0",
-        "Chrome/91.0",
-        "Safari/14.0",
-        "Firefox/89.0",
-        "Edge/91.0",
-      ];
-
-      for (const ua of userAgents) {
-        const request = new Request("http://localhost/api/auth/login", {
-          method: "POST",
-          body: JSON.stringify({ publicKey }),
-          headers: { "User-Agent": ua },
-        });
-
-        responses.push(await POST_login(request));
-      }
-
-      // Rate limiting should be based on public key, not user agent
-      const hasRateLimiting = responses.some((r) => r.status === 429);
-
-      if (!hasRateLimiting && responses.length > 3) {
-        console.warn(
-          "⚠️  Rate limiting may be bypassable via header manipulation",
-        );
-      }
-
-      expect(responses.length).toBe(userAgents.length);
-    });
-
-    it("should not be bypassed by changing nonce values", async () => {
-      const publicKey = testKeypair.publicKey();
-
-      // Get multiple challenges
-      const challenges = await Promise.all(
-        Array(10)
-          .fill(null)
-          .map(async () => {
-            const req = new Request("http://localhost/api/auth/login", {
-              method: "POST",
-              body: JSON.stringify({ publicKey }),
-            });
-            const res = await POST_login(req);
-            return res.json();
-          }),
-      );
-
-      // All should have unique nonces
-      const nonces = challenges.map((c) => c.data?.nonce);
-      const uniqueNonces = new Set(nonces);
-
-      expect(uniqueNonces.size).toBe(challenges.length);
-    });
-  });
-
-  describe("Time-based rate limit windows", () => {
-    it("should reset rate limits after time window expires", async () => {
-      const publicKey = testKeypair.publicKey();
-
-      // Make initial requests
-      for (let i = 0; i < 3; i++) {
-        const request = new Request("http://localhost/api/auth/login", {
-          method: "POST",
-          body: JSON.stringify({ publicKey }),
-        });
-        await POST_login(request);
-      }
-
-      // Wait for rate limit window to expire (typically 1 minute)
-      // In real tests, you'd mock time or use shorter windows
-      // await new Promise(resolve => setTimeout(resolve, 61000));
-
-      // Should be able to make requests again
-      const request = new Request("http://localhost/api/auth/login", {
-        method: "POST",
-        body: JSON.stringify({ publicKey }),
-      });
-
-      const response = await POST_login(request);
-
-      // Document expected behavior
-      expect(response.status).toBeLessThanOrEqual(200);
-    });
-  });
-
-  describe("Rate limiting recommendations", () => {
-    it("should document recommended rate limit configuration", () => {
-      const recommendations = {
-        loginEndpoint: {
-          window: "1 minute",
-          maxRequests: 10,
-          keyBy: "publicKey",
-        },
-        verifyEndpoint: {
-          window: "5 minutes",
-          maxRequests: 5,
-          maxFailures: 3,
-          keyBy: "publicKey",
-        },
-        globalEndpoint: {
-          window: "1 minute",
-          maxRequests: 100,
-          keyBy: "ip",
-        },
-      };
-
-      // Verify recommendations are documented
-      expect(recommendations.loginEndpoint.maxRequests).toBeLessThanOrEqual(20);
-      expect(recommendations.verifyEndpoint.maxFailures).toBeLessThanOrEqual(5);
-      expect(recommendations.globalEndpoint.maxRequests).toBeGreaterThan(0);
-
-      console.log("📋 Rate limiting recommendations:", recommendations);
-    });
-  });
-});
