+++ apps/web/__tests__/helpers/auth-helpers.ts
@@ -1,11 +1,6 @@
 import crypto from "crypto";
 
-/**
- * Generate a valid test Stellar keypair
- * Note: This is a mock function - actual Keypair should be mocked in tests
- */
 export function generateTestKeypair() {
-  // Return a mock keypair structure - actual Keypair will be mocked in tests
   return {
     publicKey: () => "GDUMMYPUBLICKEY123456789",
     sign: jest.fn().mockReturnValue(Buffer.from("mockSignature", "base64")),
@@ -13,15 +8,11 @@ export function generateTestKeypair() {
   };
 }
 
-// Generate test challenge data
-
 export function generateTestChallenge() {
   const nonce = crypto.randomBytes(32).toString("hex");
   const timestamp = Date.now();
   return { nonce, timestamp };
 }
-
-//Sign a message with a Stellar keypair
 
 export function signMessage(keypair: any, message: string): string {
   const messageBuffer = Buffer.from(message, "utf-8");
@@ -29,16 +20,10 @@ export function signMessage(keypair: any, message: string): string {
   return signature.toString("base64");
 }
 
-/**
- * Create test request body for login endpoint
- */
 export function createLoginBody(publicKey: string) {
   return { publicKey };
 }
 
-/**
- * Create test request body for verify endpoint
- */
 export function createVerifyBody(
   publicKey: string,
   signature: string,
@@ -48,23 +33,14 @@ export function createVerifyBody(
   return { publicKey, signature, nonce, timestamp };
 }
 
-/**
- * Mock expired timestamp (older than 5 minutes)
- */
 export function getExpiredTimestamp(): number {
-  return Date.now() - 6 * 60 * 1000; // 6 minutes ago
+  return Date.now() - 6 * 60 * 1000;
 }
 
-/**
- * Mock future timestamp (more than 5 minutes ahead)
- */
 export function getFutureTimestamp(): number {
-  return Date.now() + 6 * 60 * 1000; // 6 minutes from now
+  return Date.now() + 6 * 60 * 1000;
 }
 
-/**
- * Create test HTTP request with JSON body
- */
 export function createAuthRequest(
   url: string,
   body: any,
@@ -80,9 +56,6 @@ export function createAuthRequest(
   });
 }
 
-/**
- * Parse Set-Cookie header into structured data
- */
 export function parseCookie(cookieString: string | null): {
   name: string;
   value: string;
@@ -101,9 +74,6 @@ export function parseCookie(cookieString: string | null): {
   };
 }
 
-/**
- * Check if cookie has specific attribute
- */
 export function cookieHasAttribute(
   cookieString: string | null,
   attribute: string,
