--- apps/web/lib/validation/input.test.ts
@@ -1,359 +1,0 @@
-/**
- * @file input.test.ts
- * @description Test cases for input validation and sanitization
- * 
- * Run this file to verify validation works correctly:
- * npx tsx lib/validation/input.test.ts
- */
-
-import {
-  emailSchema,
-  usernameSchema,
-  passwordSchema,
-  stellarPublicKeySchema,
-  titleSchema,
-  descriptionSchema,
-  messageContentSchema,
-  searchQuerySchema,
-  validateInput,
-  sanitizeHTML,
-  sanitizePlainText,
-  containsSuspiciousPatterns,
-  validateFilename,
-  createSafeError,
-} from './input';
-
-// Color codes for terminal output
-const colors = {
-  green: '\x1b[32m',
-  red: '\x1b[31m',
-  yellow: '\x1b[33m',
-  reset: '\x1b[0m',
-};
-
-function test(name: string, fn: () => void) {
-  try {
-    fn();
-    console.log(`${colors.green}✓${colors.reset} ${name}`);
-  } catch (error) {
-    console.log(`${colors.red}✗${colors.reset} ${name}`);
-    console.error(error);
-  }
-}
-
-function assert(condition: boolean, message: string) {
-  if (!condition) {
-    throw new Error(message);
-  }
-}
-
-console.log('\n🔐 Input Validation & Sanitization Tests\n');
-
-// ──────────────────────────────────────────────────────────────
-// Email Validation Tests
-// ──────────────────────────────────────────────────────────────
-
-console.log('📧 Email Validation:');
-
-test('Valid email passes', () => {
-  const result = validateInput(emailSchema, 'user@example.com');
-  assert(result.success === true, 'Should succeed');
-  assert(result.success && result.data === 'user@example.com', 'Should return email');
-});
-
-test('Invalid email fails', () => {
-  const result = validateInput(emailSchema, 'not-an-email');
-  assert(result.success === false, 'Should fail');
-  assert(!result.success && result.errors[''] !== undefined, 'Should have error');
-});
-
-test('Email with XSS attempt is sanitized', () => {
-  const result = validateInput(emailSchema, 'test<script>alert(1)</script>@example.com');
-  assert(result.success === false, 'Should fail validation');
-});
-
-test('Email is lowercased', () => {
-  const result = validateInput(emailSchema, 'USER@EXAMPLE.COM');
-  assert(result.success && result.data === 'user@example.com', 'Should be lowercase');
-});
-
-// ──────────────────────────────────────────────────────────────
-// Username Validation Tests
-// ──────────────────────────────────────────────────────────────
-
-console.log('\n👤 Username Validation:');
-
-test('Valid username passes', () => {
-  const result = validateInput(usernameSchema, 'john_doe-123');
-  assert(result.success === true, 'Should succeed');
-});
-
-test('Username too short fails', () => {
-  const result = validateInput(usernameSchema, 'ab');
-  assert(result.success === false, 'Should fail');
-});
-
-test('Username with special chars fails', () => {
-  const result = validateInput(usernameSchema, 'user@name!');
-  assert(result.success === false, 'Should fail');
-});
-
-test('Username with spaces fails', () => {
-  const result = validateInput(usernameSchema, 'user name');
-  assert(result.success === false, 'Should fail');
-});
-
-// ──────────────────────────────────────────────────────────────
-// Password Validation Tests
-// ──────────────────────────────────────────────────────────────
-
-console.log('\n🔑 Password Validation:');
-
-test('Strong password passes', () => {
-  const result = validateInput(passwordSchema, 'MyP@ssw0rd!');
-  assert(result.success === true, 'Should succeed');
-});
-
-test('Weak password fails', () => {
-  const result = validateInput(passwordSchema, 'password');
-  assert(result.success === false, 'Should fail');
-});
-
-test('Password without uppercase fails', () => {
-  const result = validateInput(passwordSchema, 'myp@ssw0rd!');
-  assert(result.success === false, 'Should fail');
-});
-
-test('Password without special char fails', () => {
-  const result = validateInput(passwordSchema, 'MyPassw0rd');
-  assert(result.success === false, 'Should fail');
-});
-
-// ──────────────────────────────────────────────────────────────
-// Stellar Address Validation Tests
-// ──────────────────────────────────────────────────────────────
-
-console.log('\n⭐ Stellar Address Validation:');
-
-test('Valid Stellar address passes', () => {
-  const validAddress = 'GBRPYHIL2CI3FNQ4BXLFMNDLFJUNPU2HY3ZMFSHONUCEOASW7QC7OX2H';
-  const result = validateInput(stellarPublicKeySchema, validAddress);
-  assert(result.success === true, 'Should succeed');
-});
-
-test('Invalid Stellar address (wrong prefix) fails', () => {
-  const result = validateInput(stellarPublicKeySchema, 'SBRPYHIL2CI3FNQ4BXLFMNDLFJUNPU2HY3ZMFSHONUCEOASW7QC7OX2H');
-  assert(result.success === false, 'Should fail');
-});
-
-test('Invalid Stellar address (wrong length) fails', () => {
-  const result = validateInput(stellarPublicKeySchema, 'GBRPYHIL2CI3FNQ4BXLFMNDLFJUNPU2HY');
-  assert(result.success === false, 'Should fail');
-});
-
-// ──────────────────────────────────────────────────────────────
-// HTML Sanitization Tests
-// ──────────────────────────────────────────────────────────────
-
-console.log('\n🧹 HTML Sanitization:');
-
-test('Safe HTML tags are preserved', () => {
-  const input = '<p>Hello <strong>world</strong>!</p>';
-  const output = sanitizeHTML(input);
-  assert(output === input, 'Safe HTML should be preserved');
-});
-
-test('Script tags are removed', () => {
-  const input = '<p>Hello</p><script>alert("XSS")</script>';
-  const output = sanitizeHTML(input);
-  assert(!output.includes('script'), 'Script tags should be removed');
-  assert(output.includes('Hello'), 'Safe content should remain');
-});
-
-test('Event handlers are removed', () => {
-  const input = '<p onclick="alert(1)">Click me</p>';
-  const output = sanitizeHTML(input);
-  assert(!output.includes('onclick'), 'Event handlers should be removed');
-});
-
-test('Javascript protocol is removed', () => {
-  const input = '<a href="javascript:alert(1)">Link</a>';
-  const output = sanitizeHTML(input);
-  assert(!output.includes('javascript:'), 'Javascript protocol should be removed');
-});
-
-test('Plain text sanitization removes all HTML', () => {
-  const input = '<p>Hello <strong>world</strong>!</p>';
-  const output = sanitizePlainText(input);
-  assert(output === 'Hello world!', 'All HTML should be removed');
-});
-
-// ──────────────────────────────────────────────────────────────
-// Content Validation Tests
-// ──────────────────────────────────────────────────────────────
-
-console.log('\n📝 Content Validation:');
-
-test('Valid title passes', () => {
-  const result = validateInput(titleSchema, 'Beautiful Apartment Downtown');
-  assert(result.success === true, 'Should succeed');
-});
-
-test('Title with HTML is sanitized', () => {
-  const result = validateInput(titleSchema, 'Title <script>alert(1)</script>');
-  assert(result.success === true, 'Should succeed');
-  assert(result.success && !result.data.includes('script'), 'HTML should be removed');
-});
-
-test('Description allows safe HTML', () => {
-  const input = '<p>Great location with <strong>amazing</strong> views!</p>';
-  const result = validateInput(descriptionSchema, input);
-  assert(result.success === true, 'Should succeed');
-  assert(result.success && result.data.includes('<strong>'), 'Safe HTML should remain');
-});
-
-test('Description removes unsafe HTML', () => {
-  const input = '<p>Great location</p><script>alert(1)</script>';
-  const result = validateInput(descriptionSchema, input);
-  assert(result.success === true, 'Should succeed');
-  assert(result.success && !result.data.includes('script'), 'Unsafe HTML should be removed');
-});
-
-test('Message content strips all HTML', () => {
-  const result = validateInput(messageContentSchema, 'Hello <b>friend</b>!');
-  assert(result.success === true, 'Should succeed');
-  assert(result.success && result.data === 'Hello friend!', 'HTML should be stripped');
-});
-
-// ──────────────────────────────────────────────────────────────
-// Search Query Sanitization Tests
-// ──────────────────────────────────────────────────────────────
-
-console.log('\n🔍 Search Query Sanitization:');
-
-test('Normal search query passes', () => {
-  const result = validateInput(searchQuerySchema, 'apartment downtown');
-  assert(result.success === true, 'Should succeed');
-});
-
-test('SQL injection attempt is sanitized', () => {
-  const result = validateInput(searchQuerySchema, "'; DROP TABLE users; --");
-  assert(result.success === true, 'Should succeed');
-  assert(result.success && !result.data.includes(';'), 'Semicolons should be removed');
-  assert(result.success && !result.data.includes('--'), 'SQL comments should be removed');
-});
-
-test('Search query with HTML is sanitized', () => {
-  const result = validateInput(searchQuerySchema, '<script>alert(1)</script>');
-  assert(result.success === true, 'Should succeed');
-  assert(result.success && !result.data.includes('script'), 'HTML should be removed');
-});
-
-// ──────────────────────────────────────────────────────────────
-// Suspicious Pattern Detection Tests
-// ──────────────────────────────────────────────────────────────
-
-console.log('\n🚨 Suspicious Pattern Detection:');
-
-test('XSS attempts are detected', () => {
-  const patterns = [
-    '<script>alert(1)</script>',
-    'javascript:alert(1)',
-    '<img src=x onerror=alert(1)>',
-    '<iframe src="evil.com"></iframe>',
-  ];
-  patterns.forEach((pattern) => {
-    assert(containsSuspiciousPatterns(pattern), `Should detect: ${pattern}`);
-  });
-});
-
-test('SQL injection attempts are detected', () => {
-  const patterns = [
-    "'; DROP TABLE users; --",
-    "' OR '1'='1",
-    'UNION SELECT * FROM passwords',
-    'INSERT INTO users VALUES',
-    'DELETE FROM users WHERE',
-  ];
-  patterns.forEach((pattern) => {
-    assert(containsSuspiciousPatterns(pattern), `Should detect: ${pattern}`);
-  });
-});
-
-test('Path traversal attempts are detected', () => {
-  const patterns = [
-    '../../../etc/passwd',
-    '..\\..\\..\\windows\\system32',
-  ];
-  patterns.forEach((pattern) => {
-    assert(containsSuspiciousPatterns(pattern), `Should detect: ${pattern}`);
-  });
-});
-
-test('Normal input is not flagged', () => {
-  const inputs = [
-    'Hello world!',
-    'This is a normal message.',
-    'Looking for an apartment in downtown',
-    'Price: $1,200/month',
-  ];
-  inputs.forEach((input) => {
-    assert(!containsSuspiciousPatterns(input), `Should NOT flag: ${input}`);
-  });
-});
-
-// ──────────────────────────────────────────────────────────────
-// Filename Validation Tests
-// ──────────────────────────────────────────────────────────────
-
-console.log('\n📁 Filename Validation:');
-
-test('Valid filename passes', () => {
-  const result = validateFilename('document.pdf');
-  assert(result === 'document.pdf', 'Should return sanitized filename');
-});
-
-test('Filename with path traversal fails', () => {
-  const result = validateFilename('../../../etc/passwd');
-  assert(result === null, 'Should return null');
-});
-
-test('Hidden file fails', () => {
-  const result = validateFilename('.htaccess');
-  assert(result === null, 'Should return null');
-});
-
-test('Filename with invalid chars fails', () => {
-  const result = validateFilename('file<script>.pdf');
-  assert(result === null, 'Should return null');
-});
-
-// ──────────────────────────────────────────────────────────────
-// Safe Error Messages Tests
-// ──────────────────────────────────────────────────────────────
-
-console.log('\n⚠️  Safe Error Messages:');
-
-test('Error message escapes HTML', () => {
-  const error = createSafeError('<script>alert(1)</script>');
-  assert(!error.error.includes('<script>'), 'Should escape HTML');
-  assert(error.error.includes('&lt;script&gt;'), 'Should use HTML entities');
-});
-
-test('Error includes timestamp', () => {
-  const error = createSafeError('Test error');
-  assert(error.timestamp !== undefined, 'Should have timestamp');
-});
-
-test('Error includes field when provided', () => {
-  const error = createSafeError('Invalid input', 'email');
-  assert(error.field === 'email', 'Should include field');
-});
-
-// ──────────────────────────────────────────────────────────────
-// Summary
-// ──────────────────────────────────────────────────────────────
-
-console.log(`\n${colors.green}✅ All tests passed!${colors.reset}\n`);
-console.log('🔐 Input validation is working correctly.');
-console.log('🛡️  Protection against XSS, SQL injection, and other attacks is in place.\n');
