+++ apps/web/package.json
@@ -52,8 +52,8 @@
         "@testing-library/jest-dom": "^6.9.1",
         "@testing-library/react": "^16.3.2",
         "@testing-library/user-event": "^14.6.1",
-        "@types/jest": "^30.0.0",
         "@types/geojson": "^7946.0.16",
+        "@types/jest": "^30.0.0",
         "@types/jsonwebtoken": "^9.0.10",
         "@types/mapbox__point-geometry": "^0.1.4",
         "@types/node": "^20",
