+++ e2e/playwright.config.ts
@@ -6,7 +6,9 @@ export default defineConfig({
   forbidOnly: !!process.env.CI,
   retries: process.env.CI ? 2 : 0,
   workers: process.env.CI ? 1 : undefined,
-  reporter: process.env.CI ? [["html", { open: "never" }], ["github"]] : [["html"]],
+  reporter: process.env.CI
+    ? [["html", { open: "never" }], ["github"]]
+    : [["html"]],
   timeout: 30_000,
   expect: { timeout: 5_000 },
